+++ src/components/common/Collapsible.tsx
